+++ modules/AWSPowerShell/Cmdlets/GroundStation/Basic/New-GSDataflowEndpointGroup-Cmdlet.cs
@@ -53,9 +53,10 @@ namespace Amazon.PowerShell.Cmdlets.GS
         #region Parameter ContactPostPassDurationSecond
         /// <summary>
         /// <para>
-        /// <para>Amount of time, in seconds, after a contact ends for the contact to remain in a <code>POSTPASS</code>
-        /// state. A CloudWatch event is emitted when the contact enters and exits the <code>POSTPASS</code>
-        /// state.</para>
+        /// <para>Amount of time, in seconds, after a contact ends that the Ground Station Dataflow
+        /// Endpoint Group will be in a <code>POSTPASS</code> state. A Ground Station Dataflow
+        /// Endpoint Group State Change event will be emitted when the Dataflow Endpoint Group
+        /// enters and exits the <code>POSTPASS</code> state.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
@@ -66,9 +67,10 @@ namespace Amazon.PowerShell.Cmdlets.GS
         #region Parameter ContactPrePassDurationSecond
         /// <summary>
         /// <para>
-        /// <para>Amount of time, in seconds, prior to contact start for the contact to remain in a
-        /// <code>PREPASS</code> state. A CloudWatch event is emitted when the contact enters
-        /// and exits the <code>PREPASS</code> state.</para>
+        /// <para>Amount of time, in seconds, before a contact starts that the Ground Station Dataflow
+        /// Endpoint Group will be in a <code>PREPASS</code> state. A Ground Station Dataflow
+        /// Endpoint Group State Change event will be emitted when the Dataflow Endpoint Group
+        /// enters and exits the <code>PREPASS</code> state.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]