+++ modules/AWSPowerShell/Cmdlets/ManagedBlockchain/Basic/New-MBCMember-Cmdlet.cs
@@ -51,7 +51,7 @@ namespace Amazon.PowerShell.Cmdlets.MBC
         /// <summary>
         /// <para>
         /// <para>The password for the member's initial administrative user. The <code>AdminPassword</code>
-        /// must be at least eight characters long and no more than 32 characters. It must contain
+        /// must be at least 8 characters long and no more than 32 characters. It must contain
         /// at least one uppercase letter, one lowercase letter, and one digit. It cannot have
         /// a single quotation mark (‘), a double quotation marks (“), a forward slash(/), a backward
         /// slash(\), @, or a space.</para>